move CLI front-end tests from .cli.tests to .ide.tests bundle
