+++ tests/org.eclipse.n4js.integration.tests/src/org/eclipse/n4js/integration/tests/cli/frontend/FrontendLspTest.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.cli.frontend.tests;
+package org.eclipse.n4js.integration.tests.cli.frontend;
 
 import static org.junit.Assert.assertEquals;
